emploi-matching/requirements.txt
requests
pandas
beautifulsoup4
python-dotenv
pymongo
sqlalchemy
fastapi
faker
